Electron: add a skip when electron.app installed
Sometimes the network is not good enough to download electron from aws
service especially in China mainland. And the `download` module doesn't
show the progress bar in real-time. That might not be a good experience.

In this commit, just added a function check if the electron.app has been
copied to `outputDir`(download electron.app from other high-speed tool),
then this installation can be skiped.